lib/Sagent/blog/requirements.txt
Flask==2.3.3
markdown==3.4.4
PyYAML==6.0.1
python-frontmatter==1.0.0